+++ src/vote.cpp
@@ -2,32 +2,26 @@
 void wps::vote( const eosio::name voter, const eosio::name proposal_name, const eosio::name vote )
 {
     require_auth( voter );
-    const eosio::name ram_payer = get_self();
 
-    // validate proposal
-    auto proposals_itr = _proposals.find( proposal_name.value );
-    check( proposals_itr != _proposals.end(), "[proposal_name] does not exists");
-    check( proposals_itr->start_voting_period <= current_time_point(), "[proposal_name] has not yet started");
-    check( proposals_itr->end > current_time_point(), "[proposal_name] has ended");
-
-    // validate vote
-    auto votes_itr = _votes.find( proposal_name.value );
-    check( votes_itr != _votes.end(), "[proposal_name] votes does not exist");
-    check( vote == "yes"_n || vote == "no"_n || vote == "abstain"_n, "[vote] invalid (ex: yes/no/abstain)");
+    // validate proposal if eligible to vote on
+    check_proposal_can_vote( proposal_name );
 
     // cannot vote during completed voting period phase
     check_voting_period_completed();
 
-    // update vote
-    _votes.modify( votes_itr, ram_payer, [&]( auto& vote_row ) {
-        check(vote_row.votes[voter] != vote, "[vote] has not been modified");
-        vote_row.votes[voter] = vote;
+    // update `votes` table
+    update_vote( voter, proposal_name, vote );
 
-        // re-caculate votes
-        _proposals.modify( proposals_itr, ram_payer, [&]( auto& proposals_row ) {
-            proposals_row.total_net_votes = calculate_total_net_votes( vote_row.votes );
-        });
-    });
+    // update `proposals::eligible` field for all active proposals
+    update_eligible_proposals();
+}
+
+void wps::check_proposal_can_vote( const eosio::name proposal_name )
+{
+    auto proposals_itr = _proposals.find( proposal_name.value );
+    check( proposals_itr != _proposals.end(), "[proposal_name] does not exists");
+    check( proposals_itr->start_voting_period <= current_time_point(), "[proposal_name] has not yet started");
+    check( proposals_itr->end > current_time_point(), "[proposal_name] has ended");
 }
 
 int16_t wps::calculate_total_net_votes( const std::map<eosio::name, eosio::name> votes )
@@ -39,4 +33,78 @@ int16_t wps::calculate_total_net_votes( const std::map<eosio::name, eosio::name>
         else if (vote == "no"_n) total_net_votes -= 1;
     }
     return total_net_votes;
+}
+
+void wps::update_vote( const eosio::name voter, const eosio::name proposal_name, const eosio::name vote )
+{
+    // validate vote
+    auto votes_itr = _votes.find( proposal_name.value );
+    auto proposals_itr = _proposals.find( proposal_name.value );
+
+    check( votes_itr != _votes.end(), "[proposal_name] votes does not exist");
+    check( vote == "yes"_n || vote == "no"_n || vote == "abstain"_n, "[vote] invalid (ex: yes/no/abstain)");
+
+    // update vote
+    _votes.modify( votes_itr, same_payer, [&]( auto& vote_row ) {
+        check(vote_row.votes[voter] != vote, "[vote] has not been modified");
+        vote_row.votes[voter] = vote;
+
+        // re-caculate votes
+        _proposals.modify( proposals_itr, same_payer, [&]( auto& proposals_row ) {
+            proposals_row.total_net_votes = calculate_total_net_votes( vote_row.votes );
+        });
+    });
+}
+
+void wps::update_eligible_proposals()
+{
+    // settings
+    auto settings = _settings.get();
+    auto state = _state.get();
+
+    // containers
+    eosio::asset total_payout = asset{ 0, symbol{ "EOS", 4 }};
+
+    // filter out min voting threshold proposals
+    std::map<int16_t, std::set<eosio::name>> proposals = sort_proposals_by_net_votes( "active"_n );
+
+    // iterate proposals from highest to lowest net votes
+    for ( auto itr = proposals.rbegin(); itr != proposals.rend(); ++itr ) {
+
+        // iterate over proposals
+        for ( auto proposal_name : itr->second ) {
+            // proposal variables
+            auto proposal_itr = _proposals.find( proposal_name.value );
+            const int16_t total_net_votes = itr->first;
+            const eosio::name proposer = proposal_itr->proposer;
+            const eosio::asset monthly_budget = proposal_itr->monthly_budget;
+
+            // min requirements for payouts
+            const bool is_min_vote_margin = total_net_votes >= settings.vote_margin;
+            const bool is_enough_budget = (total_payout + monthly_budget) <= settings.max_monthly_budget;
+
+            // set eligible of proposal (true/false)
+            _proposals.modify( proposal_itr, same_payer, [&]( auto& row ) {
+                if ( is_enough_budget && is_min_vote_margin ) {
+                    total_payout += monthly_budget;
+                    row.eligible = true;
+                } else {
+                    row.eligible = false;
+                }
+            });
+        }
+    }
+}
+
+std::map<int16_t, std::set<eosio::name>> wps::sort_proposals_by_net_votes( const eosio::name status )
+{
+    auto index = _proposals.get_index<"bystatus"_n>();
+
+    // add proposals to ordered map
+    std::map<int16_t, std::set<eosio::name>> proposals;
+
+    for ( auto itr = index.lower_bound(status.value); itr != index.upper_bound(status.value); ++itr ) {
+        proposals[itr->total_net_votes].insert(itr->proposal_name);
+    }
+    return proposals;
 }